requirements.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
coverage==7.4.1
decorator==5.1.1
dj-database-url==2.1.0
Django==5.0.2
django-environ==0.11.2
django-heroku==0.3.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-extensions==3.1.3
executing==2.0.1
idna==3.6
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.21.0
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pytest>=6.0.0
python-decouple==3.5
pluggy==1.4.0
psycopg2==2.9.9
pytest==8.0.0
pytest-cov==4.1.0
pytest-django==4.8.0
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.2.0
wcwidth==0.2.13
whitenoise==6.6.0